requirements.txt
langchain
langchain-openai
langchainhub
mido
flask
flask-cors
git-python
rich
hydra-core
autoflake
music21
omegaconf